split SpaceStatisticsAPI to separate file space_statistics.py
